// src/types/ClientMessageType.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PresentationOperation } from "./PresentationOperation";
import type { PresentationPresenceUpdate } from "./PresentationPresenceUpdate";

export type ClientMessageType = { "type": "JoinRoom", "data": { room_id: string, } } | { "type": "LeaveRoom" } | { "type": "UpdatePresence", "data": { presence: PresentationPresenceUpdate, } } | { "type": "UpdateStorage", "data": { operations: Array<PresentationOperation>, } };